DB-9327 allow SSQ in update statement to be flattened, and disable BatchOnce optimization
